implemented full functionality for nominees, categories, school, and voting

+++ src/main/java/rigor/io/paragala/voter/details/DataController.java
@@ -1,6 +1,5 @@
 package rigor.io.paragala.voter.details;
 
-import jdk.nashorn.internal.parser.Token;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -9,11 +8,7 @@ import rigor.io.paragala.voter.nominees.Nominee;
 import rigor.io.paragala.voter.nominees.NomineeRepository;
 import rigor.io.paragala.voter.token.TokenService;
 
-import javax.xml.ws.Response;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.List;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @RestController
@@ -40,16 +35,26 @@ public class DataController {
     this.nomineeRepository = nomineeRepository;
   }
 
+  @DeleteMapping("/schools/{school}")
+  public ResponseEntity<?> deleteSchool(@RequestParam(required = false) String token,
+                                        @PathVariable String school) {
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
+
+    schoolRepository.deleteByName(school);
+    return new ResponseEntity<>("ok", HttpStatus.ACCEPTED);
+  }
+
   @GetMapping("/schools")
   public ResponseEntity<?> getSchools(@RequestParam(required = false) String token) {
-//    if (!tokenService.isValid(token))
-//      return ResponseHub.defaultUnauthorizedResponse();
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
 
     List<School> schools = schoolRepository.findAll();
     return new ResponseEntity<>(
         schools
             .stream()
-            .map(School::getSchool)
+            .map(School::getName)
             .collect(Collectors.toList())
             .toArray(new String[schools.size()]),
         HttpStatus.OK
@@ -59,19 +64,34 @@ public class DataController {
   @PostMapping("/schools")
   public ResponseEntity<?> addSchools(@RequestParam(required = false) String token,
                                       @RequestBody String[] schools) {
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
+
     schoolRepository.deleteAll();
     List<School> schoolList = Arrays.stream(schools).map(School::new).collect(Collectors.toList());
     return new ResponseEntity<>(schoolRepository.saveAll(schoolList), HttpStatus.OK);
   }
 
+  @DeleteMapping("/categories/{category}")
+  public ResponseEntity<?> deleteCategory(@RequestParam(required = false) String token,
+                                          @PathVariable String category) {
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
+
+    categoryRepository.deleteByKategory(category);
+    return new ResponseEntity<>("eh e heh", HttpStatus.ACCEPTED);
+  }
+
   @GetMapping("/categories")
   public ResponseEntity<?> getCategories(@RequestParam(required = false) String token) {
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
 
     List<Category> categories = categoryRepository.findAll();
     return new ResponseEntity<>(
         categories
             .stream()
-            .map(Category::getCategory)
+            .map(Category::getKategory)
             .collect(Collectors.toList())
             .toArray(new String[categories.size()]),
         HttpStatus.OK
@@ -81,6 +101,9 @@ public class DataController {
   @PostMapping("/categories")
   public ResponseEntity<?> addCategories(@RequestParam(required = false) String token,
                                          @RequestBody String[] categories) {
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
+
     categoryRepository.deleteAll();
     List<Category> categoryList = Arrays.stream(categories).map(Category::new).collect(Collectors.toList());
     Iterable<Category> extractedBecauseMeNoLikeSeeingYellowLines = categoryRepository.saveAll(categoryList);
@@ -88,8 +111,9 @@ public class DataController {
   }
 
   @GetMapping("/nominees")
-  public ResponseEntity<?> getNominees(@RequestParam(required = false) String token,
-                                       @RequestBody List<String> nominees) {
+  public ResponseEntity<?> getNominees(@RequestParam(required = false) String token) {
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
 
     return new ResponseEntity<>(nomineeRepository.findAll(), HttpStatus.OK);
   }
@@ -97,17 +121,33 @@ public class DataController {
   @PostMapping("/nominees")
   public ResponseEntity<?> addNominees(@RequestParam(required = false) String token,
                                        @RequestBody List<Nominee> nominees) {
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
 
     nomineeRepository.deleteAll();
     return new ResponseEntity<>(nomineeRepository.saveAll(nominees), HttpStatus.OK);
+  }
+
+  @DeleteMapping("/nominees")
+  public ResponseEntity<?> deleteNominee(@RequestParam(required = false) String token,
+                                         @RequestBody Map<String, String> data) {
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
+
+    String title = data.get("title");
+    String category = data.get("kategory");
+    nomineeRepository.deleteByTitleAndCategory(title, category);
+    return new ResponseEntity<>("ayye", HttpStatus.ACCEPTED);
   }
 
 
   // DEFAULTS
   @GetMapping("/defaults/categories")
   public ResponseEntity<?> defaultCategories(@RequestParam(required = false) String token) {
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
 
-//    categoryRepository.deleteAll();
+    //    categoryRepository.deleteAll();
     return new ResponseEntity<>(
         new HashSet<>(Arrays.asList(categories)),
         HttpStatus.OK
@@ -116,6 +156,8 @@ public class DataController {
 
   @GetMapping("/defaults/nominees")
   public ResponseEntity<?> defaultNominees(@RequestParam(required = false) String token) {
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
 
     nomineeRepository.deleteAll();
     return new ResponseEntity<>(nomineeRepository.saveAll(populate()), HttpStatus.OK);
@@ -123,8 +165,10 @@ public class DataController {
 
   @GetMapping("/defaults/schools")
   public ResponseEntity<?> defaultSchools(@RequestParam(required = false) String token) {
+    if (!tokenService.isValid(token))
+      return ResponseHub.defaultUnauthorizedResponse();
 
-    return new ResponseEntity<>(new HashSet<String>(){{
+    return new ResponseEntity<>(new HashSet<String>() {{
       add("Holy Angel University");
       add("Angeles University Foundation");
       add("Mabalacat City College");
@@ -144,6 +188,5 @@ public class DataController {
     }
     return nominees;
   }
-
 
 }
